requirements.txt
# a_unet==0.0.16
# botocore
# braceexpand==0.1.7
# diffusers==0.29.2
# torch==2.4.1
# jsonargparse==4.32.1
# librosa==0.10.1
# numpy==1.26.1
# omegaconf==2.3.0
# packaging==24.1
# pandas==2.2.3
# Pillow==10.4.0
# progressbar33==2.4
# pytorch_lightning==2.4.0
# PyYAML==6.0.1
# regex==2023.10.3
# sagemaker==2.232.1
# scikit_learn==1.3.2
# soundfile==0.12.1
# timm==1.0.9
# torchaudio==2.4.1
# torchlibrosa==0.1.0
# torchvision==0.19.1
# tqdm==4.65.0
# transformers==4.34.1
# huggingface_hub==0.23.4
# tokenizers==0.14.1
# wandb==0.18.1
# webdataset==0.2.100
# wget==3.2
# jsonargparse[signatures]>=4.27.7
# lightning-utilities==0.9.0
# torchmetrics==1.3.2
# python-dateutil
# scipy==1.11.3
# joblib==1.3.2


a-unet==0.0.16
absl-py==2.0.0
aiofiles==23.2.1
aiobotocore
aiohttp
aiohappyeyeballs
aioitertools
aiosignal
multidict
yarl
altair==5.1.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argbind==0.3.7
attrs==23.1.0
audiomentations==0.34.1
audioread==3.0.1
av==10.0.0
awscli==1.33.9
blis==0.7.11
boto3==1.35.14
botocore==1.35.22
braceexpand==0.1.7
build==0.10.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
confection==0.1.3
contourpy==1.1.1
coverage==7.4.3
cycler==0.12.1
cymem==2.0.8
data==0.4
data-science-types==0.2.23
diffusers==0.29.2
dill==0.3.8
docopt==0.6.2
docstring-parser==0.15
docutils==0.16
dora-search==0.1.12
einops==0.7.0
fastapi==0.104.0
ffmpy==0.3.1
filelock==3.12.4
fire==0.5.0
flashy==0.0.2
flatten-dict==0.4.2
fonttools==4.43.1
fsspec==2023.9.2
ftfy==6.2.0
funcsigs==1.0.2
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.2
h11==0.14.0
h5py==3.11.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.23.4
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipython-genutils==0.2.0
Jinja2==3.1.2
joblib==1.3.2
jsonargparse==4.26.1
jsonschema==4.19.1
julius==0.2.7
kiwisolver==1.4.5
lameenc==1.6.1
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
lightning-utilities==0.9.0
linformer==0.2.1
llvmlite==0.41.1
mdurl==0.1.2
mido==1.3.2
mir-eval==0.7
mock==4.0.3
mpmath==1.3.0
msgpack==1.0.7
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.2
num2words==0.5.13
numba==0.58.1
numpy==1.26.1
oauthlib==3.2.2
omegaconf==2.3.0
openunmix==1.2.1
orjson==3.9.9
pandas==2.1.1
pathos==0.3.2
pathtools==0.1.2
pathy==0.10.3
Pillow==10.1.0
platformdirs==4.3.6
plotly==5.19.0
pluggy==1.4.0
pooch==1.7.0
pox==0.3.4
ppft==1.7.6.8
preshed==3.0.9
primePy==1.3
product-key-memory==0.2.10
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==1.10.13
pydub==0.25.1
pyloudnorm==0.1.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrubberband==0.3.0
PySoundFile==0.9.0.post1
pystoi==0.3.3
pytest==8.0.2
pytest-cov==4.1.0
python-multipart==0.0.6
pytorch-lightning==2.1.0
pytz==2023.3.post1
PyYAML==6.0.1
randomname==0.2.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.8.1
rsa==4.7.2
s3transfer==0.10.1
safetensors==0.4.0
schema==0.7.7
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
shellingham==1.5.4
shutilwhich==1.1.0
smart-open==6.4.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
soundfile==0.12.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
submitit==1.5.0
sympy==1.12
tblib==3.0.0
tempdir==0.7.1
tenacity==8.2.3
termcolor==2.3.0
thinc==8.1.12
threadpoolctl==3.2.0
timm==0.9.16
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch-stoi==0.1.2
# torch==2.1.0
# torchaudio==2.1.0
# torchlibrosa==0.1.0
torchmetrics==1.3.2
tqdm==4.66.1
transformers==4.34.1
treetable==0.2.5
triton==2.1.0
typer==0.7.0
typeshed-client==2.4.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.23.2
wandb==0.15.12
wasabi==1.1.2
webdataset==0.2.86
websockets==11.0.3
Werkzeug==3.0.1
wget==3.2
yarg==0.1.9
typing-extensions
python-dateutil
torchlibrosa==0.1.0
jmespath
sniffio
# s3fs
async-timeout
docker-pycreds
rich

